backend_test_villeda/requirements.txt
aiohttp==3.7.3
amqp==5.0.5
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-model-utils==4.1.1
djangorestframework==3.12.2
drf-yasg==1.20.0
gevent==21.1.2
greenlet==1.0.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
multidict==5.1.0
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
prompt-toolkit==3.0.16
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
slack-sdk==3.4.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
